+++ src/main/java/com/graduatioinProject/sensorMonitoring/formerData/electricCurrent/ElectricCurrentService.java
@@ -1,13 +1,10 @@
-package com.graduatioinProject.sensorMonitoring.node.formerData.electricCurrent;
+package com.graduatioinProject.sensorMonitoring.formerData.electricCurrent;
 
-import com.graduatioinProject.sensorMonitoring.node.formerData.temperature.Temperature;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
